Add better memory tracking

+++ src/main/java/net/irisshaders/batchedentityrendering/impl/MemoryTrackingBuffer.java
@@ -1,9 +1,9 @@
 package net.irisshaders.batchedentityrendering.impl;
 
 public interface MemoryTrackingBuffer {
-	int getAllocatedSize();
+	long getAllocatedSize();
 
-	int getUsedSize();
+	long getUsedSize();
 
 	void freeAndDeleteBuffer();
 }
